chore : 캐러셀 컴포넌트화
